+++ src/azure-cli/azure/cli/command_modules/storage/_help.py
@@ -19,14 +19,14 @@ short-summary: Manage storage accounts.
 
 helps['storage account blob-service-properties'] = """
 type: group
-short-summary: Manage blob service properties for storage account.
+short-summary: Manage the properties of a storage account's blob service.
 """
 
 helps['storage account blob-service-properties show'] = """
 type: command
-short-summary: Show the properties of a storage account's Blob service.
+short-summary: Show the properties of a storage account's blob service.
 long-summary: >
-    Show the properties of a storage account's Blob service, including
+    Show the properties of a storage account's blob service, including
     properties for Storage Analytics and CORS (Cross-Origin Resource
     Sharing) rules.
 examples:
@@ -36,19 +36,17 @@ examples:
 
 helps['storage account blob-service-properties update'] = """
 type: command
-short-summary: Update the properties of a storage account's Blob service.
+short-summary: Update the properties of a storage account's blob service.
 long-summary: >
-    Update the properties of a storage account's Blob service, including
+    Update the properties of a storage account's blob service, including
     properties for Storage Analytics and CORS (Cross-Origin Resource
-    Sharing) rules.
+    Sharing) rules. But currently we only support enabling or disabling change feed for a storage account's blob service.
 parameters:
   - name: --enable-change-feed
     short-summary: 'Indicate whether change feed event logging is enabled. If it is true, you enable the storage account to begin capturing changes. The default value is true. You can see more details in https://docs.microsoft.com/en-us/azure/storage/blobs/storage-blob-change-feed?tabs=azure-portal#register-by-using-azure-cli'
 examples:
   - name: Enable the change feed for the storage account 'MyStorageAccount' in resource group 'MyResourceGroup'.
     text: az storage account blob-service-properties update --enable-change-feed true -n MyStorageAccount -g MyResourceGroup
-  - name: Disable the change feed for the storage account 'MyStorageAccount' in resource group 'MyResourceGroup'.
-    text: az storage account blob-service-properties update --enable-change-feed false -n MyStorageAccount -g MyResourceGroup
 """
 
 helps['storage account create'] = """